requirements.txt
asgiref==3.4.1
attrs==21.2.0
bootstrap4==0.1.0
Brotli==1.0.9
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
configobj==5.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
Cython==0.29.17
defusedxml==0.7.1
Django==3.0.5
django-allauth==0.44.0
django-cors-headers==3.7.0
django-extensions==3.1.3
django-filter==2.2.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.2.0
django-storages==1.11.1
djangorestframework==3.12.4
docopt==0.6.2
drf-yasg==1.20.0
environ==1.0
factory-boy==3.2.0
Faker==8.9.1
fcm-django==0.3.10
future==0.18.2
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.6.1
inflection==0.5.1
iniconfig==1.1.1
ipaddr==2.2.0
itypes==1.2.0
Jinja2==3.0.1
lxml==4.5.0
MarkupSafe==2.0.1
mock==4.0.3
module==0.0.4
oauthlib==3.1.1
ordereddict==1.1
packaging==21.0
pipreq==0.4
pipreqs==0.4.10
pluggy==0.13.1
protobuf==3.17.3
psycopg2==2.9.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyfcm==1.4.9
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-gnupg==0.4.7
python-ldap==3.3.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
systematic==4.8.7
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.2.0
wincertstore==0.2
yarg==0.1.9
zipp==3.5.0